+++ src/js/input-range/input-range.jsx
@@ -27,6 +27,7 @@ export default class InputRange extends React.Component {
       ariaControls: PropTypes.string,
       classNames: PropTypes.objectOf(PropTypes.string),
       disabled: PropTypes.bool,
+      draggableTrack: PropTypes.bool,
       formatLabel: PropTypes.func,
       maxValue: rangePropType,
       minValue: rangePropType,
@@ -98,6 +99,12 @@ export default class InputRange extends React.Component {
      * @type {?Component}
      */
     this.trackNode = null;
+
+    /**
+     * @private
+     * @type {bool}
+     */
+    this.isSliderDragging = false;
   }
 
   /**
@@ -376,8 +383,44 @@ export default class InputRange extends React.Component {
     }
 
     const position = valueTransformer.getPositionFromEvent(event, this.getTrackClientRect());
-
+    this.isSliderDragging = true;
     requestAnimationFrame(() => this.updatePosition(key, position));
+  }
+
+  /**
+   * Handle any "mousemove" event received by the track
+   * @private
+   * @param {SyntheticEvent} event
+   * @return {void}
+   */
+  @autobind
+  handleTrackDrag(event, prevEvent) {
+    if (this.props.disabled || !this.props.draggableTrack || this.isSliderDragging) {
+      return;
+    }
+
+    const {
+      maxValue,
+      minValue,
+      value: { max, min },
+    } = this.props;
+
+    const position = valueTransformer.getPositionFromEvent(event, this.getTrackClientRect());
+    const value = valueTransformer.getValueFromPosition(position, minValue, maxValue, this.getTrackClientRect());
+    const stepValue = valueTransformer.getStepValueFromValue(value, this.props.step);
+
+    const prevPosition = valueTransformer.getPositionFromEvent(prevEvent, this.getTrackClientRect());
+    const prevValue = valueTransformer.getValueFromPosition(prevPosition, minValue, maxValue, this.getTrackClientRect());
+    const prevStepValue = valueTransformer.getStepValueFromValue(prevValue, this.props.step);
+
+    const offset = prevStepValue - stepValue;
+
+    const transformedValues = {
+      min: min - offset,
+      max: max - offset,
+    };
+
+    this.updateValues(transformedValues);
   }
 
   /**
@@ -424,9 +467,20 @@ export default class InputRange extends React.Component {
       return;
     }
 
+    const {
+      maxValue,
+      minValue,
+      value: { max, min },
+    } = this.props;
+
     event.preventDefault();
 
-    this.updatePosition(this.getKeyByPosition(position), position);
+    const value = valueTransformer.getValueFromPosition(position, minValue, maxValue, this.getTrackClientRect());
+    const stepValue = valueTransformer.getStepValueFromValue(value, this.props.step);
+
+    if (!this.props.draggableTrack || stepValue > max || stepValue < min) {
+      this.updatePosition(this.getKeyByPosition(position), position);
+    }
   }
 
   /**
@@ -452,6 +506,10 @@ export default class InputRange extends React.Component {
    */
   @autobind
   handleInteractionEnd() {
+    if (this.isSliderDragging) {
+      this.isSliderDragging = false;
+    }
+
     if (!this.props.onChangeComplete || !isDefined(this.startValue)) {
       return;
     }
@@ -673,14 +731,16 @@ export default class InputRange extends React.Component {
 
         <Track
           classNames={this.props.classNames}
+          draggableTrack={this.props.draggableTrack}
           ref={(trackNode) => { this.trackNode = trackNode; }}
           percentages={percentages}
           suggestedPercentages={suggestedPercentages}
-          onTrackMouseDown={this.handleTrackMouseDown}
           withActive={this.props.withActive}
           isMultiValue={this.isMultiValue()}
           withError={this.hasSingleValueError()}
-          errorPercentages={errorPercentages}>
+          errorPercentages={errorPercentages}
+          onTrackDrag={this.handleTrackDrag}
+          onTrackMouseDown={this.handleTrackMouseDown}>
 
           {this.renderSliders()}
         </Track>