+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+<a name="1.2.1"></a>
+## [1.2.1](https://github.com/davidchin/react-input-range/compare/v1.2.0...v1.2.1) (2017-07-14)
+
+
+### Bug Fixes
+
+* Always set slider dragging back to false ([#98](https://github.com/davidchin/react-input-range/issues/98)) ([d22fa26](https://github.com/davidchin/react-input-range/commit/d22fa26))
+
+
+
+<a name="1.2.0"></a>
+# [1.2.0](https://github.com/davidchin/react-input-range/compare/v1.1.5...v1.2.0) (2017-07-09)
+
+
+### Features
+
+* Add track dragging functionality ([#91](https://github.com/davidchin/react-input-range/issues/91)) ([4a8ca26](https://github.com/davidchin/react-input-range/commit/4a8ca26))
+
+
+
+<a name="1.1.5"></a>
+## [1.1.5](https://github.com/davidchin/react-input-range/compare/v1.1.4...v1.1.5) (2017-07-09)
+
+
+### Bug Fixes
+
+* Fix Typescript definition file ([4935745](https://github.com/davidchin/react-input-range/commit/4935745))
+
+
+
 <a name="1.1.4"></a>
 ## [1.1.4](https://github.com/davidchin/react-input-range/compare/v1.1.3...v1.1.4) (2017-05-20)